+++ src/view/src/components/UI/Buttons/CloseTransparentButton.vue
