+++ src/view/src/components/UI/index.ts
@@ -1,10 +1,10 @@
-import DarkButton from "@/components/UI/DarkButton.vue"
+import DarkButton from "@/components/UI/Buttons/DarkButton.vue"
 import LoadingIcon from "@/components/UI/LoadingIcon.vue"
-import TransparentButton from "@/components/UI/TransparentButton.vue"
+import TransparentButton from "@/components/UI/Buttons/TransparentButton.vue"
 import TransparentInput from "@/components/UI/TransparentInput.vue"
 import ErrorNotice from "@/components/UI/ErrorNotice.vue"
-import CloseTransparentButton from "@/components/UI/CloseTransparentButton.vue"
-import TransparantButtonNB from "@/components/UI/TransparantButtonNB.vue"
+import CloseTransparentButton from "@/components/UI/Buttons/CloseTransparentButton.vue"
+import TransparantButtonNB from "@/components/UI/Buttons/TransparantButtonNB.vue"
 
 export default [
     DarkButton,
